[teleport-update] status subcommand (#49308)
* status

* cleanup

* comments

* cleanup output by removing optional fields

* rebase fix
