+++ lib/autoupdate/agent/testdata/TestUpdater_Disable/already_disabled.golden
@@ -2,11 +2,8 @@ version: v1
 kind: update_config
 spec:
     proxy: ""
-    group: ""
-    url_template: ""
     enabled: false
     pinned: false
 status:
     active_version: ""
     backup_version: ""
-    skip_version: ""
