+++ lib/autoupdate/agent/testdata/TestUpdater_Install/config_from_user.golden
@@ -9,4 +9,3 @@ spec:
 status:
     active_version: new-version
     backup_version: old-version
-    skip_version: ""
