+++ lib/autoupdate/agent/testdata/TestUpdater_Install/insecure_URL.golden
@@ -2,11 +2,9 @@ version: v1
 kind: update_config
 spec:
     proxy: ""
-    group: ""
     url_template: http://example.com
     enabled: false
     pinned: false
 status:
     active_version: ""
     backup_version: ""
-    skip_version: ""
