+++ lib/autoupdate/agent/testdata/TestUpdater_Install/invalid_metadata.golden
@@ -2,11 +2,8 @@ version: ""
 kind: ""
 spec:
     proxy: ""
-    group: ""
-    url_template: ""
     enabled: false
     pinned: false
 status:
     active_version: ""
     backup_version: ""
-    skip_version: ""
