+++ lib/autoupdate/agent/testdata/TestUpdater_Update/insecure_URL.golden
@@ -2,11 +2,9 @@ version: v1
 kind: update_config
 spec:
     proxy: localhost
-    group: ""
     url_template: http://example.com
     enabled: true
     pinned: false
 status:
     active_version: ""
     backup_version: ""
-    skip_version: ""
