+++ lib/autoupdate/agent/testdata/TestUpdater_Update/install_error.golden
@@ -2,11 +2,8 @@ version: v1
 kind: update_config
 spec:
     proxy: localhost
-    group: ""
-    url_template: ""
     enabled: true
     pinned: false
 status:
     active_version: ""
     backup_version: ""
-    skip_version: ""
