+++ lib/autoupdate/agent/testdata/TestUpdater_Update/pinned_version.golden
@@ -2,11 +2,9 @@ version: v1
 kind: update_config
 spec:
     proxy: localhost
-    group: ""
     url_template: https://example.com
     enabled: true
     pinned: true
 status:
     active_version: old-version
     backup_version: backup-version
-    skip_version: ""
