+++ lib/autoupdate/agent/testdata/TestUpdater_Update/updates_enabled_during_window.golden
@@ -9,4 +9,3 @@ spec:
 status:
     active_version: 16.3.0
     backup_version: old-version
-    skip_version: ""
