+++ lib/autoupdate/agent/testdata/TestUpdater_Update/version_already_installed_outside_of_window.golden
@@ -2,11 +2,9 @@ version: v1
 kind: update_config
 spec:
     proxy: localhost
-    group: ""
     url_template: https://example.com
     enabled: true
     pinned: false
 status:
     active_version: 16.3.0
     backup_version: ""
-    skip_version: ""
